[Decomp][GameObject] Cohesion split: extract Update/Use + destructible to GameObject*.cpp
Mirror of mangosthree #267 + #268 (bundled, same source file). Pure
cohesion split of GameObject.cpp with no behaviour change:

- GameObjectUpdate.cpp        - the per-tick GameObject::Update state
  machine (carries the file-local Dire Maul fixed-trap enum it uses).
- GameObjectUse.cpp           - GameObject::Use interaction dispatch.
- GameObjectDestructible.cpp  - capture-point slider/tick and destructible
  building damage/rebuild/force-health.

GameObject.h is unchanged. CMake file(GLOB Object/*.cpp) auto-compiles the
new TUs. WotLK has no SetInUse (Cata-only).
