+++ src/game/Object/GameObject.cpp
@@ -56,12 +56,6 @@
 #include "LuaEngine.h"
 #endif /* ENABLE_ELUNA */
 
-enum
-{
-    GO_DIRE_MAUL_FIXED_TRAP = 179512,
-    NPC_SLIPKIK_GUARD = 14323
-};
-
 /**
  * @brief Creates a game object instance with default runtime state.
  */
@@ -320,390 +314,6 @@ bool GameObject::Create(uint32 guidlow, uint32 name_id, Map* map, uint32 phaseMa
     return true;
 }
 
-/**
- * @brief Updates game object state, timers, loot state, and AI.
- *
- * @param update_diff The elapsed AI update time in milliseconds.
- * @param p_time The elapsed world time step in milliseconds.
- */
-void GameObject::Update(uint32 update_diff, uint32 p_time)
-{
-    if (GetObjectGuid().IsMOTransport())
-    {
-        //((Transport*)this)->Update(p_time);
-        return;
-    }
-
-    // Used by Eluna
-#ifdef ENABLE_ELUNA
-    if (Eluna* e = GetEluna())
-    {
-        e->UpdateAI(this, update_diff);
-    }
-#endif /* ENABLE_ELUNA */
-
-    switch (m_lootState)
-    {
-        case GO_NOT_READY:
-        {
-            switch (GetGoType())
-            {
-                case GAMEOBJECT_TYPE_TRAP:                  // Initialized delayed to be able to use GetOwner()
-                {
-                    // Arming Time for GAMEOBJECT_TYPE_TRAP (6)
-                    Unit* owner = GetOwner();
-                    if (owner && owner->IsInCombat())
-                    {
-                        m_cooldownTime = time(NULL) + GetGOInfo()->trap.startDelay;
-                    }
-                    m_lootState = GO_READY;
-                    break;
-                }
-                case GAMEOBJECT_TYPE_FISHINGNODE:           // Keep not ready for some delay
-                {
-                    // fishing code (bobber ready)
-                    if (time(NULL) > m_respawnTime - FISHING_BOBBER_READY_TIME)
-                    {
-                        // splash bobber (bobber ready now)
-                        Unit* caster = GetOwner();
-                        if (caster && caster->GetTypeId() == TYPEID_PLAYER)
-                        {
-                            SetGoState(GO_STATE_ACTIVE);
-                            // SetUInt32Value(GAMEOBJECT_FLAGS, GO_FLAG_NODESPAWN);
-
-                            SendForcedObjectUpdate();
-
-                            SendGameObjectCustomAnim(GetObjectGuid());
-                        }
-
-                        m_lootState = GO_READY;             // can be successfully open with some chance
-                    }
-                    break;
-                }
-                case GAMEOBJECT_TYPE_CHEST:
-                {
-                    if (m_goInfo->chest.chestRestockTime)
-                    {
-                        if (m_reStockTimer != 0)
-                        {
-                            if (m_reStockTimer <= time(nullptr))
-                            {
-                                m_reStockTimer = 0;
-                                m_lootState = GO_READY;
-                                loot.clear();
-                                MarkFlagUpdateForClient(GAMEOBJECT_DYNAMIC); // GAMEOBJECT_DYN_FLAGS);
-                            }
-                        }
-                        else
-                        {
-                            m_lootState = GO_READY;
-                        }
-                        return;
-                    }
-                    m_lootState = GO_READY;
-                }
-                default:
-                    break;
-            }
-            break;
-        }
-        case GO_READY:
-        {
-            if (m_respawnTime > 0)                          // timer on
-            {
-                if (m_respawnTime <= time(NULL))            // timer expired
-                {
-                    m_respawnTime = 0;
-                    ClearAllUsesData();
-
-                    switch (GetGoType())
-                    {
-                        case GAMEOBJECT_TYPE_FISHINGNODE:   // can't fish now
-                        {
-                            Unit* caster = GetOwner();
-                            if (caster && caster->GetTypeId() == TYPEID_PLAYER)
-                            {
-                                caster->FinishSpell(CURRENT_CHANNELED_SPELL);
-
-                                WorldPacket data(SMSG_FISH_NOT_HOOKED, 0);
-                                ((Player*)caster)->GetSession()->SendPacket(&data);
-                            }
-                            // can be deleted
-                            m_lootState = GO_JUST_DEACTIVATED;
-                            return;
-                        }
-                        case GAMEOBJECT_TYPE_DOOR:
-                        case GAMEOBJECT_TYPE_BUTTON:
-                            // we need to open doors if they are closed (add there another condition if this code breaks some usage, but it need to be here for battlegrounds)
-                            if (GetGoState() != GO_STATE_READY)
-                            {
-                                ResetDoorOrButton();
-                            }
-                            // flags in AB are type_button and we need to add them here so no break!
-                        default:
-                            if (!m_spawnedByDefault)        // despawn timer
-                            {
-                                // can be despawned or destroyed
-                                SetLootState(GO_JUST_DEACTIVATED);
-                                // Remove Wild-Summoned GO on timer expire
-                                if (!HasStaticDBSpawnData())
-                                {
-                                    if (Unit* owner = GetOwner())
-                                    {
-                                        owner->RemoveGameObject(this, false);
-                                    }
-                                    Delete();
-                                }
-                                return;
-                            }
-
-                            // respawn timer
-                            GetMap()->Add(this);
-                            break;
-                    }
-                }
-            }
-
-            if (isSpawned())
-            {
-                // traps can have time and can not have
-                GameObjectInfo const* goInfo = GetGOInfo();
-                uint32 max_charges = goInfo->GetCharges();
-
-                if (goInfo->type == GAMEOBJECT_TYPE_TRAP)   // traps
-                {
-                    if (m_cooldownTime >= time(NULL))
-                    {
-                        return;
-                    }
-
-                    // FIXME: this is activation radius (in different casting radius that must be selected from spell data)
-                    // TODO: move activated state code (cast itself) to GO_ACTIVATED, in this place only check activating and set state
-                    float radius = float(goInfo->trap.radius);
-                    if (!radius)
-                    {
-                        if (goInfo->trap.cooldown != 3)     // cast in other case (at some triggering/linked go/etc explicit call)
-                        {
-                            return;
-                        }
-                        else
-                        {
-                            if (m_respawnTime > 0)
-                            {
-                                break;
-                            }
-
-                            // battlegrounds gameobjects has data2 == 0 && data5 == 3
-                            radius = float(goInfo->trap.cooldown);
-                        }
-                    }
-
-                    // Should trap trigger?
-                    Unit* enemy = NULL;                     // pointer to appropriate target if found any
-                    MaNGOS::AnyUnfriendlyUnitInObjectRangeCheck u_check(this, radius);
-                    MaNGOS::UnitSearcher<MaNGOS::AnyUnfriendlyUnitInObjectRangeCheck> checker(enemy, u_check);
-                    Cell::VisitAllObjects(this, checker, radius);
-                    if (enemy)
-                    {
-                        bool useTrap = true;
-                        // prevent use if GO entry is "Fixed Trap" and target is not SLIKIK
-                        if (GetEntry() == GO_DIRE_MAUL_FIXED_TRAP && enemy->GetEntry() != NPC_SLIPKIK_GUARD)
-                        {
-                            useTrap = false;
-                        }
-
-                        if (useTrap)
-                        {
-                            Use(enemy);
-                        }
-                    }
-                }
-
-                // Only despawn object if there are charges to "consume"
-                // it means (all GO with charges = 0 in DB should never be despawned)
-                // Check : https://www.getmangos.eu/wiki/referenceinfo/dbinfo/mangosdb/mangoszeroworlddb/gameobject_template-r1047
-                // for more information about charges field in db depending on object type
-                if (max_charges > 0 && m_useTimes >= max_charges)
-                {
-                    m_useTimes = 0;
-                    SetLootState(GO_JUST_DEACTIVATED);  // can be despawned or destroyed
-                }
-            }
-            break;
-        }
-        case GO_ACTIVATED:
-        {
-            switch (GetGoType())
-            {
-                case GAMEOBJECT_TYPE_DOOR:
-                case GAMEOBJECT_TYPE_BUTTON:
-                    if (GetGOInfo()->GetAutoCloseTime() && (m_cooldownTime < time(NULL)))
-                    {
-                        ResetDoorOrButton();
-                    }
-                    break;
-                case GAMEOBJECT_TYPE_CHEST:
-                    if (true)
-                    {
-                        if (!loot.empty())
-                        {
-                            m_despawnTimer = time(nullptr) + 5 * MINUTE; // TODO:: need to add a define?
-                        }
-                        else if (m_despawnTimer != 0 && m_despawnTimer <= time(nullptr))
-                        {
-                            m_lootState = GO_JUST_DEACTIVATED;
-                        }
-
-                        // TODO : Missing Loot::Update() method found in CMangos
-                    }
-                    break;
-                case GAMEOBJECT_TYPE_TRAP:
-                    if (m_rearmTimer == 0)
-                    {
-                        m_rearmTimer = time(nullptr) + GetRespawnDelay();
-                        SetGoState(GO_STATE_ACTIVE_ALTERNATIVE);
-                    }
-
-                    if (m_rearmTimer < time(nullptr))
-                    {
-                        SetGoState(GO_STATE_READY);
-                        m_lootState = GO_READY;
-                        m_rearmTimer = 0;
-                    }
-                    break;
-                case GAMEOBJECT_TYPE_GOOBER:
-                    if (m_cooldownTime < time(NULL))
-                    {
-                        RemoveFlag(GAMEOBJECT_FLAGS, GO_FLAG_IN_USE);
-
-                        SetLootState(GO_JUST_DEACTIVATED);
-                        m_cooldownTime = 0;
-                    }
-                    break;
-                case GAMEOBJECT_TYPE_CAPTURE_POINT:
-                    m_captureTimer += p_time;
-                    if (m_captureTimer >= 5000)
-                    {
-                        TickCapturePoint();
-                        m_captureTimer -= 5000;
-                    }
-                    break;
-                default:
-                    break;
-            }
-            break;
-        }
-        case GO_JUST_DEACTIVATED:
-        {
-            switch (GetGoType())
-            {
-                case GAMEOBJECT_TYPE_GOOBER:
-                    // if gameobject should cast spell, then this, but some GOs (type = 10) should be destroyed
-                    if (uint32 spellId = GetGOInfo()->goober.spellId)
-                    {
-                        for (GuidSet::const_iterator itr = m_UniqueUsers.begin(); itr != m_UniqueUsers.end(); ++itr)
-                        {
-                            if (Player* owner = GetMap()->GetPlayer(*itr))
-                            {
-                                owner->CastSpell(owner, spellId, false, NULL, NULL, GetObjectGuid());
-                            }
-                        }
-
-                        ClearAllUsesData();
-                    }
-
-                    SetGoState(GO_STATE_READY);
-
-                    // any return here in case battleground traps
-                    break;
-
-                case GAMEOBJECT_TYPE_CAPTURE_POINT:
-                    // remove capturing players because slider wont be displayed if capture point is being locked
-                    for (GuidSet::const_iterator itr = m_UniqueUsers.begin(); itr != m_UniqueUsers.end(); ++itr)
-                    {
-                        if (Player* owner = GetMap()->GetPlayer(*itr))
-                        {
-                            owner->SendUpdateWorldState(GetGOInfo()->capturePoint.worldState1, WORLD_STATE_REMOVE);
-                        }
-                    }
-
-                    m_UniqueUsers.clear();
-                    SetLootState(GO_READY);
-                    return; // SetLootState and return because go is treated as "burning flag" due to GetGoAnimProgress() being 100 and would be removed on the client
-                default:
-                    break;
-            }
-
-            // Remove wild summoned after use
-            if (!HasStaticDBSpawnData() && (!GetSpellId() || GetGOInfo()->GetDespawnPossibility() || GetGOInfo()->IsDespawnAtAction()))
-            {
-                if (Unit* owner = GetOwner())
-                {
-                    owner->RemoveGameObject(this, false);
-                }
-                Delete();
-                return;
-            }
-
-            // burning flags in some battlegrounds, if you find better condition, just add it
-            if (GetGOInfo()->IsDespawnAtAction() || GetGoAnimProgress() > 0)
-            {
-                SendObjectDeSpawnAnim(GetObjectGuid());
-                // reset flags
-                if (GetMap()->Instanceable())
-                {
-                    // In Instances GO_FLAG_LOCKED, GO_FLAG_INTERACT_COND or GO_FLAG_NO_INTERACT are not changed
-                    uint32 currentLockOrInteractFlags = GetUInt32Value(GAMEOBJECT_FLAGS) & (GO_FLAG_LOCKED | GO_FLAG_INTERACT_COND | GO_FLAG_NO_INTERACT);
-                    SetUInt32Value(GAMEOBJECT_FLAGS, (GetGOInfo()->flags & ~(GO_FLAG_LOCKED | GO_FLAG_INTERACT_COND | GO_FLAG_NO_INTERACT)) | currentLockOrInteractFlags);
-                }
-                else
-                {
-                    SetUInt32Value(GAMEOBJECT_FLAGS, GetGOInfo()->flags);
-                }
-            }
-
-            loot.clear();
-            SetLootRecipient(NULL);
-            SetLootState(GO_READY);
-
-            if (!m_respawnDelayTime)
-            {
-                return;
-            }
-
-            // since pool system can fail to roll unspawned object, this one can remain spawned, so must set respawn nevertheless
-            m_respawnTime = m_spawnedByDefault ? time(NULL) + m_respawnDelayTime : 0;
-
-            // if option not set then object will be saved at grid unload
-            if (sWorld.getConfig(CONFIG_BOOL_SAVE_RESPAWN_TIME_IMMEDIATELY))
-            {
-                SaveRespawnTime();
-            }
-
-            // if part of pool, let pool system schedule new spawn instead of just scheduling respawn
-            if (uint16 poolid = sPoolMgr.IsPartOfAPool<GameObject>(GetGUIDLow()))
-            {
-                sPoolMgr.UpdatePool<GameObject>(*GetMap()->GetPersistentState(), poolid, GetGUIDLow());
-            }
-
-            // can be not in world at pool despawn
-            if (IsInWorld())
-            {
-                UpdateObjectVisibility();
-            }
-
-            break;
-        }
-    }
-
-    if (AI())
-    {
-        // do not allow the AI to be changed during update
-        m_AI_locked = true;
-        AI()->UpdateAI(update_diff);   // AI not react good at real update delays (while freeze in non-active part of map)
-        m_AI_locked = false;
-    }
-}
 
 /**
  * @brief Refreshes the game object spawn state on the map.
@@ -1452,796 +1062,6 @@ void GameObject::SwitchDoorOrButton(bool activate, bool alternative /* = false *
     }
 }
 
-/**
- * @brief Handles use interaction for this game object.
- *
- * @param user The unit using the object.
- */
-void GameObject::Use(Unit* user)
-{
-    // user must be provided
-    MANGOS_ASSERT(user || PrintEntryError("GameObject::Use (without user)"));
-
-    // by default spell caster is user
-    Unit* spellCaster = user;
-    uint32 spellId = 0;
-    bool triggered = false;
-
-    // test only for exist cooldown data (cooldown timer used for door/buttons reset that not have use cooldown)
-    if (uint32 cooldown = GetGOInfo()->GetCooldown())
-    {
-        if (m_cooldownTime > sWorld.GetGameTime())
-        {
-            return;
-        }
-
-        m_cooldownTime = sWorld.GetGameTime() + cooldown;
-    }
-
-    bool scriptReturnValue = user->GetTypeId() == TYPEID_PLAYER && sScriptMgr.OnGameObjectUse((Player*)user, this);
-    if (!scriptReturnValue)
-    {
-        GetMap()->ScriptsStart(DBS_ON_GOT_USE, GetEntry(), spellCaster, this);
-    }
-
-    switch (GetGoType())
-    {
-        case GAMEOBJECT_TYPE_DOOR:                          // 0
-        {
-            // doors never really despawn, only reset to default state/flags
-            UseDoorOrButton();
-
-            // activate script
-            if (!scriptReturnValue)
-            {
-                GetMap()->ScriptsStart(DBS_ON_GO_USE, GetGUIDLow(), spellCaster, this);
-            }
-            return;
-        }
-        case GAMEOBJECT_TYPE_BUTTON:                        // 1
-        {
-            // buttons never really despawn, only reset to default state/flags
-            UseDoorOrButton();
-
-            TriggerLinkedGameObject(user);
-
-            // activate script
-            if (!scriptReturnValue)
-            {
-                GetMap()->ScriptsStart(DBS_ON_GO_USE, GetGUIDLow(), spellCaster, this);
-            }
-
-            return;
-        }
-        case GAMEOBJECT_TYPE_QUESTGIVER:                    // 2
-        {
-            if (user->GetTypeId() != TYPEID_PLAYER)
-            {
-                return;
-            }
-
-            Player* player = (Player*)user;
-
-            if (!sScriptMgr.OnGossipHello(player, this))
-            {
-                player->PrepareGossipMenu(this, GetGOInfo()->questgiver.gossipID);
-                player->SendPreparedGossip(this);
-            }
-
-            return;
-        }
-        case GAMEOBJECT_TYPE_CHEST:                         // 3
-        {
-            if (user->GetTypeId() != TYPEID_PLAYER)
-            {
-                return;
-            }
-
-            TriggerLinkedGameObject(user);
-
-            // TODO: possible must be moved to loot release (in different from linked triggering)
-            if (GetGOInfo()->chest.eventId)
-            {
-                DEBUG_LOG("Chest ScriptStart id %u for %s (opened by %s)", GetGOInfo()->chest.eventId, GetGuidStr().c_str(), user->GetGuidStr().c_str());
-                StartEvents_Event(GetMap(), GetGOInfo()->chest.eventId, user, this);
-            }
-
-            return;
-        }
-        case GAMEOBJECT_TYPE_GENERIC:                       // 5
-        {
-            if (scriptReturnValue)
-            {
-                return;
-            }
-
-            // No known way to exclude some - only different approach is to select despawnable GOs by Entry
-            SetLootState(GO_JUST_DEACTIVATED);
-            return;
-        }
-        case GAMEOBJECT_TYPE_TRAP:                          // 6
-        {
-            if (scriptReturnValue)
-            {
-                return;
-            }
-
-            Unit* owner = GetOwner();
-            Unit* caster = owner ? owner : user;
-
-            GameObjectInfo const* goInfo = GetGOInfo();
-            float radius = float(goInfo->trap.radius);
-            bool IsBattleGroundTrap = !radius && goInfo->trap.cooldown == 3 && m_respawnTime == 0;
-
-            // FIXME: when GO casting will be implemented trap must cast spell to target
-            if (goInfo->trap.spellId)
-            {
-                caster->CastSpell(user, goInfo->trap.spellId, true, NULL, NULL, GetObjectGuid());
-            }
-            // use template cooldown if provided
-            m_cooldownTime = time(NULL) + (goInfo->trap.cooldown ? goInfo->trap.cooldown : uint32(4));
-
-            // count charges
-            if (goInfo->trap.charges > 0)
-            {
-                AddUse();
-            }
-
-            if (IsBattleGroundTrap && user->GetTypeId() == TYPEID_PLAYER)
-            {
-                // BattleGround gameobjects case
-                if (BattleGround* bg = ((Player*)user)->GetBattleGround())
-                {
-                    bg->HandleTriggerBuff(GetObjectGuid());
-                }
-            }
-
-            // TODO: all traps can be activated, also those without spell.
-            // Some may have have animation and/or are expected to despawn.
-
-            // TODO: Improve this when more information is available, currently these traps are known that must send the anim (Onyxia/ Heigan Fissures/ Trap in DireMaul)
-            if (GetDisplayId() == 4392 || GetDisplayId() == 4472 || GetDisplayId() == 4491 || GetDisplayId() == 6785 || GetDisplayId() == 3073 || GetDisplayId() == 7998)
-            {
-                SendGameObjectCustomAnim(GetObjectGuid());
-            }
-
-            if (!scriptReturnValue && user->GetTypeId() == TYPEID_UNIT)
-            {
-                sScriptMgr.OnGameObjectUse(user, this);
-            }
-
-            // TODO: Despawning of traps? (Also related to code in ::Update)
-            return;
-        }
-        case GAMEOBJECT_TYPE_CHAIR:                         // 7 Sitting: Wooden bench, chairs
-        {
-            GameObjectInfo const* info = GetGOInfo();
-            if (!info)
-            {
-                return;
-            }
-
-            if (user->GetTypeId() != TYPEID_PLAYER)
-            {
-                return;
-            }
-
-            Player* player = (Player*)user;
-
-            // a chair may have n slots. we have to calculate their positions and teleport the player to the nearest one
-
-            // check if the db is sane
-            if (info->chair.slots > 0)
-            {
-                float lowestDist = DEFAULT_VISIBILITY_DISTANCE;
-
-                float x_lowest = GetPositionX();
-                float y_lowest = GetPositionY();
-
-                // the object orientation + 1/2 pi
-                // every slot will be on that straight line
-                float orthogonalOrientation = GetOrientation() + M_PI_F * 0.5f;
-                // find nearest slot
-                for (uint32 i = 0; i < info->chair.slots; ++i)
-                {
-                    // the distance between this slot and the center of the go - imagine a 1D space
-                    float relativeDistance = (info->size * i) - (info->size * (info->chair.slots - 1) / 2.0f);
-
-                    float x_i = GetPositionX() + relativeDistance * cos(orthogonalOrientation);
-                    float y_i = GetPositionY() + relativeDistance * sin(orthogonalOrientation);
-
-                    // calculate the distance between the player and this slot
-                    float thisDistance = player->GetDistance2d(x_i, y_i);
-
-                    /* debug code. It will spawn a npc on each slot to visualize them.
-                    Creature* helper = player->SummonCreature(14496, x_i, y_i, GetPositionZ(), GetOrientation(), TEMPSUMMON_TIMED_OR_DEAD_DESPAWN, 10000);
-                    std::ostringstream output;
-                    output << i << ": thisDist: " << thisDistance;
-                    helper->MonsterSay(output.str().c_str(), LANG_UNIVERSAL);
-                    */
-
-                    if (thisDistance <= lowestDist)
-                    {
-                        lowestDist = thisDistance;
-                        x_lowest = x_i;
-                        y_lowest = y_i;
-                    }
-                }
-                player->TeleportTo(GetMapId(), x_lowest, y_lowest, GetPositionZ(), GetOrientation(), TELE_TO_NOT_LEAVE_TRANSPORT | TELE_TO_NOT_LEAVE_COMBAT | TELE_TO_NOT_UNSUMMON_PET);
-            }
-            else
-            {
-                // fallback, will always work
-                player->TeleportTo(GetMapId(), GetPositionX(), GetPositionY(), GetPositionZ(), GetOrientation(), TELE_TO_NOT_LEAVE_TRANSPORT | TELE_TO_NOT_LEAVE_COMBAT | TELE_TO_NOT_UNSUMMON_PET);
-            }
-            player->SetStandState(UNIT_STAND_STATE_SIT_LOW_CHAIR + info->chair.height);
-            return;
-        }
-        case GAMEOBJECT_TYPE_SPELL_FOCUS:                   // 8
-        {
-            TriggerLinkedGameObject(user);
-
-            // some may be activated in addition? Conditions for this? (ex: entry 181616)
-            return;
-        }
-        case GAMEOBJECT_TYPE_GOOBER:                        // 10
-        {
-            // Handle OutdoorPvP use cases
-            // Note: this may be also handled by DB spell scripts in the future, when the world state manager is implemented
-            if (user->GetTypeId() == TYPEID_PLAYER)
-            {
-                Player* player = (Player*)user;
-                if (OutdoorPvP* outdoorPvP = sOutdoorPvPMgr.GetScript(player->GetCachedZoneId()))
-                {
-                    outdoorPvP->HandleGameObjectUse(player, this);
-                }
-            }
-
-            GameObjectInfo const* info = GetGOInfo();
-
-            TriggerLinkedGameObject(user);
-
-            SetFlag(GAMEOBJECT_FLAGS, GO_FLAG_IN_USE);
-            SetLootState(GO_ACTIVATED);
-
-            // this appear to be ok, however others exist in addition to this that should have custom (ex: 190510, 188692, 187389)
-            if (info->goober.customAnim)
-            {
-                SendGameObjectCustomAnim(GetObjectGuid());
-            }
-            else
-            {
-                SetGoState(GO_STATE_ACTIVE);
-            }
-
-            m_cooldownTime = time(NULL) + info->GetAutoCloseTime();
-
-            if (user->GetTypeId() == TYPEID_PLAYER)
-            {
-                Player* player = (Player*)user;
-
-                if (info->goober.pageId)                    // show page...
-                {
-                    WorldPacket data(SMSG_GAMEOBJECT_PAGETEXT, 8);
-                    data << ObjectGuid(GetObjectGuid());
-                    player->GetSession()->SendPacket(&data);
-                }
-                else if (info->goober.gossipID)             // ...or gossip, if page does not exist
-                {
-                    if (!sScriptMgr.OnGossipHello(player, this))
-                    {
-                        player->PrepareGossipMenu(this, info->goober.gossipID);
-                        player->SendPreparedGossip(this);
-                    }
-                }
-
-                if (info->goober.eventId)
-                {
-                    DEBUG_FILTER_LOG(LOG_FILTER_AI_AND_MOVEGENSS, "Goober ScriptStart id %u for %s (Used by %s).", info->goober.eventId, GetGuidStr().c_str(), player->GetGuidStr().c_str());
-                    StartEvents_Event(GetMap(), info->goober.eventId, player, this);
-                }
-
-                // possible quest objective for active quests
-                if (info->goober.questId && sObjectMgr.GetQuestTemplate(info->goober.questId))
-                {
-                    // Quest require to be active for GO using
-                    if (player->GetQuestStatus(info->goober.questId) != QUEST_STATUS_INCOMPLETE)
-                    {
-                        break;
-                    }
-                }
-
-                player->RewardPlayerAndGroupAtCast(this);
-            }
-
-            // activate script
-            if (!scriptReturnValue)
-            {
-                GetMap()->ScriptsStart(DBS_ON_GO_USE, GetGUIDLow(), spellCaster, this);
-            }
-            else
-            {
-                return;
-            }
-
-            // cast this spell later if provided
-            spellId = info->goober.spellId;
-
-            // database may contain a dummy spell, so it need replacement by actually existing
-            switch (spellId)
-            {
-                case 34448: spellId = 26566; break;
-                case 34452: spellId = 26572; break;
-                case 37639: spellId = 36326; break;
-                case 45367: spellId = 45371; break;
-                case 45370: spellId = 45368; break;
-            }
-
-            break;
-        }
-        case GAMEOBJECT_TYPE_CAMERA:                        // 13
-        {
-            GameObjectInfo const* info = GetGOInfo();
-            if (!info)
-            {
-                return;
-            }
-
-            if (user->GetTypeId() != TYPEID_PLAYER)
-            {
-                return;
-            }
-
-            Player* player = (Player*)user;
-
-            if (info->camera.cinematicId)
-            {
-                player->SendCinematicStart(info->camera.cinematicId);
-            }
-
-            if (info->camera.eventID)
-            {
-                StartEvents_Event(GetMap(), info->camera.eventID, player, this);
-            }
-
-            return;
-        }
-        case GAMEOBJECT_TYPE_FISHINGNODE:                   // 17 fishing bobber
-        {
-            if (user->GetTypeId() != TYPEID_PLAYER)
-            {
-                return;
-            }
-
-            Player* player = (Player*)user;
-
-            if (player->GetObjectGuid() != GetOwnerGuid())
-            {
-                return;
-            }
-
-            switch (getLootState())
-            {
-                case GO_READY:                              // ready for loot
-                {
-                    // 1) skill must be >= base_zone_skill
-                    // 2) if skill == base_zone_skill => 5% chance
-                    // 3) chance is linear dependence from (base_zone_skill-skill)
-
-                    uint32 zone, subzone;
-                    GetZoneAndAreaId(zone, subzone);
-
-                    int32 zone_skill = sObjectMgr.GetFishingBaseSkillLevel(subzone);
-                    if (!zone_skill)
-                    {
-                        zone_skill = sObjectMgr.GetFishingBaseSkillLevel(zone);
-                    }
-
-                    // provide error, no fishable zone or area should be 0
-                    if (!zone_skill)
-                    {
-                        sLog.outErrorDb("Fishable areaId %u are not properly defined in `skill_fishing_base_level`.", subzone);
-                    }
-
-                    int32 skill = player->GetSkillValue(SKILL_FISHING);
-                    int32 chance = skill - zone_skill + 5;
-                    int32 roll = irand(1, 100);
-
-                    DEBUG_LOG("Fishing check (skill: %i zone min skill: %i chance %i roll: %i", skill, zone_skill, chance, roll);
-
-                    // normal chance
-                    bool success = skill >= zone_skill && chance >= roll;
-                    GameObject* fishingHole = NULL;
-
-                    // overwrite fail in case fishhole if allowed (after 3.3.0)
-                    if (!success)
-                    {
-                        if (!sWorld.getConfig(CONFIG_BOOL_SKILL_FAIL_POSSIBLE_FISHINGPOOL))
-                        {
-                            // TODO: find reasonable value for fishing hole search
-                            fishingHole = LookupFishingHoleAround(20.0f + CONTACT_DISTANCE);
-                            if (fishingHole)
-                            {
-                                success = true;
-                            }
-                        }
-                    }
-                    // just search fishhole for success case
-                    else
-                        // TODO: find reasonable value for fishing hole search
-                    {
-                        fishingHole = LookupFishingHoleAround(20.0f + CONTACT_DISTANCE);
-                    }
-
-                    if (success || sWorld.getConfig(CONFIG_BOOL_SKILL_FAIL_GAIN_FISHING))
-                    {
-                        player->UpdateFishingSkill();
-                    }
-
-                    // fish catch or fail and junk allowed (after 3.1.0)
-                    if (success || sWorld.getConfig(CONFIG_BOOL_SKILL_FAIL_LOOT_FISHING))
-                    {
-                        // prevent removing GO at spell cancel
-                        player->RemoveGameObject(this, false);
-                        SetOwnerGuid(player->GetObjectGuid());
-
-                        if (fishingHole)                    // will set at success only
-                        {
-                            fishingHole->Use(player);
-                            SetLootState(GO_JUST_DEACTIVATED);
-                        }
-                        else
-                        {
-                            player->SendLoot(GetObjectGuid(), success ? LOOT_FISHING : LOOT_FISHING_FAIL);
-                        }
-                    }
-                    else
-                    {
-                        // fish escaped, can be deleted now
-                        SetLootState(GO_JUST_DEACTIVATED);
-
-                        WorldPacket data(SMSG_FISH_ESCAPED, 0);
-                        player->GetSession()->SendPacket(&data);
-                    }
-                    break;
-                }
-                case GO_JUST_DEACTIVATED:                   // nothing to do, will be deleted at next update
-                    break;
-                default:
-                {
-                    SetLootState(GO_JUST_DEACTIVATED);
-
-                    WorldPacket data(SMSG_FISH_NOT_HOOKED, 0);
-                    player->GetSession()->SendPacket(&data);
-                    break;
-                }
-            }
-
-            player->FinishSpell(CURRENT_CHANNELED_SPELL);
-            return;
-        }
-        case GAMEOBJECT_TYPE_SUMMONING_RITUAL:              // 18
-        {
-            if (user->GetTypeId() != TYPEID_PLAYER)
-            {
-                return;
-            }
-
-            Player* player = (Player*)user;
-
-            Unit* owner = GetOwner();
-
-            GameObjectInfo const* info = GetGOInfo();
-
-            if (owner)
-            {
-                if (owner->GetTypeId() != TYPEID_PLAYER)
-                {
-                    return;
-                }
-
-                // accept only use by player from same group as owner, excluding owner itself (unique use already added in spell effect)
-                if (player == (Player*)owner || (info->summoningRitual.castersGrouped && !player->IsInSameRaidWith(((Player*)owner))))
-                {
-                    return;
-                }
-
-                // expect owner to already be channeling, so if not...
-                if (!owner->GetCurrentSpell(CURRENT_CHANNELED_SPELL))
-                {
-                    return;
-                }
-
-                // in case summoning ritual caster is GO creator
-                spellCaster = owner;
-            }
-            else
-            {
-                if (m_firstUser && player->GetObjectGuid() != m_firstUser && info->summoningRitual.castersGrouped)
-                {
-                    if (Group* group = player->GetGroup())
-                    {
-                        if (!group->IsMember(m_firstUser))
-                        {
-                            return;
-                        }
-                    }
-                    else
-                    {
-                        return;
-                    }
-                }
-
-                spellCaster = player;
-            }
-
-            AddUniqueUse(player);
-
-            if (info->summoningRitual.animSpell)
-            {
-                player->CastSpell(player, info->summoningRitual.animSpell, true);
-
-                // for this case, summoningRitual.spellId is always triggered
-                triggered = true;
-            }
-
-            // full amount unique participants including original summoner, need more
-            if (GetUniqueUseCount() < info->summoningRitual.reqParticipants)
-            {
-                return;
-            }
-
-            // owner is first user for non-wild GO objects, if it offline value already set to current user
-            if (!GetOwnerGuid())
-                if (Player* firstUser = GetMap()->GetPlayer(m_firstUser))
-                {
-                    spellCaster = firstUser;
-                }
-
-            spellId = info->summoningRitual.spellId;
-
-            if (spellId == 62330)                           // GO store nonexistent spell, replace by expected
-            {
-                spellId = 61993;
-            }
-
-            // spell have reagent and mana cost but it not expected use its
-            // it triggered spell in fact casted at currently channeled GO
-            triggered = true;
-
-            // finish owners spell
-            if (owner)
-            {
-                owner->FinishSpell(CURRENT_CHANNELED_SPELL);
-            }
-
-            // can be deleted now, if
-            if (!info->summoningRitual.ritualPersistent)
-            {
-                SetLootState(GO_JUST_DEACTIVATED);
-            }
-            // reset ritual for this GO
-            else
-            {
-                ClearAllUsesData();
-            }
-
-            // go to end function to spell casting
-            break;
-        }
-        case GAMEOBJECT_TYPE_SPELLCASTER:                   // 22
-        {
-            SetUInt32Value(GAMEOBJECT_FLAGS, GO_FLAG_LOCKED);
-
-            GameObjectInfo const* info = GetGOInfo();
-            if (!info)
-            {
-                return;
-            }
-
-            if (info->spellcaster.partyOnly)
-            {
-                Unit* caster = GetOwner();
-                if (!caster || caster->GetTypeId() != TYPEID_PLAYER)
-                {
-                    return;
-                }
-
-                if (user->GetTypeId() != TYPEID_PLAYER || !((Player*)user)->IsInSameRaidWith((Player*)caster))
-                {
-                    return;
-                }
-            }
-
-            spellId = info->spellcaster.spellId;
-
-            AddUse();
-            break;
-        }
-        case GAMEOBJECT_TYPE_MEETINGSTONE:                  // 23
-        {
-            GameObjectInfo const* info = GetGOInfo();
-
-            if (user->GetTypeId() != TYPEID_PLAYER)
-            {
-                return;
-            }
-
-            Player* player = (Player*)user;
-
-            Player* targetPlayer = sObjectAccessor.FindPlayer(player->GetSelectionGuid());
-
-            // accept only use by player from same group for caster except caster itself
-            if (!targetPlayer || targetPlayer == player || !targetPlayer->IsInSameGroupWith(player))
-            {
-                return;
-            }
-
-            // required lvl checks!
-            uint8 level = player->getLevel();
-            if (level < info->meetingstone.minLevel || level > info->meetingstone.maxLevel)
-            {
-                return;
-            }
-
-            level = targetPlayer->getLevel();
-            if (level < info->meetingstone.minLevel || level > info->meetingstone.maxLevel)
-            {
-                return;
-            }
-
-            if (info->id == 194097)
-            {
-                spellId = 61994;                            // Ritual of Summoning
-            }
-            else
-            {
-                spellId = 59782;                            // Summoning Stone Effect
-            }
-
-            break;
-        }
-        case GAMEOBJECT_TYPE_FLAGSTAND:                     // 24
-        {
-            if (user->GetTypeId() != TYPEID_PLAYER)
-            {
-                return;
-            }
-
-            Player* player = (Player*)user;
-
-            if (player->CanUseBattleGroundObject())
-            {
-                // in battleground check
-                BattleGround* bg = player->GetBattleGround();
-                if (!bg)
-                {
-                    return;
-                }
-                // BG flag click
-                // AB:
-                // 15001
-                // 15002
-                // 15003
-                // 15004
-                // 15005
-                bg->EventPlayerClickedOnFlag(player, this);
-                return;                                     // we don't need to delete flag ... it is despawned!
-            }
-            break;
-        }
-        case GAMEOBJECT_TYPE_FISHINGHOLE:                   // 25
-        {
-            if (user->GetTypeId() != TYPEID_PLAYER)
-            {
-                return;
-            }
-
-            Player* player = (Player*)user;
-
-            player->SendLoot(GetObjectGuid(), LOOT_FISHINGHOLE);
-            player->UpdateAchievementCriteria(ACHIEVEMENT_CRITERIA_TYPE_FISH_IN_GAMEOBJECT, GetGOInfo()->id);
-            return;
-        }
-        case GAMEOBJECT_TYPE_FLAGDROP:                      // 26
-        {
-            if (user->GetTypeId() != TYPEID_PLAYER)
-            {
-                return;
-            }
-
-            Player* player = (Player*)user;
-
-            if (player->CanUseBattleGroundObject())
-            {
-                // in battleground check
-                BattleGround* bg = player->GetBattleGround();
-                if (!bg)
-                {
-                    return;
-                }
-                // BG flag dropped
-                // WS:
-                // 179785 - Silverwing Flag
-                // 179786 - Warsong Flag
-                // EotS:
-                // 184142 - Netherstorm Flag
-                GameObjectInfo const* info = GetGOInfo();
-                if (info)
-                {
-                    switch (info->id)
-                    {
-                        case 179785:                        // Silverwing Flag
-                        case 179786:                        // Warsong Flag
-                            // check if it's correct bg
-                            if (bg->GetTypeID(true) == BATTLEGROUND_WS)
-                            {
-                                bg->EventPlayerClickedOnFlag(player, this);
-                            }
-                            break;
-                        case 184142:                        // Netherstorm Flag
-                            if (bg->GetTypeID(true) == BATTLEGROUND_EY)
-                            {
-                                bg->EventPlayerClickedOnFlag(player, this);
-                            }
-                            break;
-                    }
-                }
-                // this cause to call return, all flags must be deleted here!!
-                spellId = 0;
-                Delete();
-            }
-            break;
-        }
-        case GAMEOBJECT_TYPE_BARBER_CHAIR:                  // 32
-        {
-            GameObjectInfo const* info = GetGOInfo();
-            if (!info)
-            {
-                return;
-            }
-
-            if (user->GetTypeId() != TYPEID_PLAYER)
-            {
-                return;
-            }
-
-            Player* player = (Player*)user;
-
-            // fallback, will always work
-            player->TeleportTo(GetMapId(), GetPositionX(), GetPositionY(), GetPositionZ(), GetOrientation(), TELE_TO_NOT_LEAVE_TRANSPORT | TELE_TO_NOT_LEAVE_COMBAT | TELE_TO_NOT_UNSUMMON_PET);
-
-            WorldPacket data(SMSG_ENABLE_BARBER_SHOP, 0);
-            player->GetSession()->SendPacket(&data);
-
-            player->SetStandState(UNIT_STAND_STATE_SIT_LOW_CHAIR + info->barberChair.chairheight);
-            return;
-        }
-        default:
-            sLog.outError("GameObject::Use unhandled GameObject type %u (entry %u).", GetGoType(), GetEntry());
-            return;
-    }
-
-    if (!spellId)
-    {
-        return;
-    }
-
-    SpellEntry const* spellInfo = sSpellStore.LookupEntry(spellId);
-    if (!spellInfo)
-    {
-        sLog.outError("WORLD: unknown spell id %u at use action for gameobject (Entry: %u GoType: %u )", spellId, GetEntry(), GetGoType());
-        return;
-    }
-
-    Spell* spell = new Spell(spellCaster, spellInfo, triggered, GetObjectGuid());
-
-    // spell target is user of GO
-    SpellCastTargets targets;
-    targets.setUnitTarget(user);
-
-    spell->prepare(&targets);
-}
 
 // overwrite WorldObject function for proper name localization
 /**
@@ -2786,428 +1606,10 @@ bool GameObject::HasStaticDBSpawnData() const
     return sObjectMgr.GetGOData(GetGUIDLow()) != NULL;
 }
 
-/**
- * @brief Sets the capture point slider and derived state.
- *
- * @param value The slider value.
- * @param isLocked true if the capture point is locked.
- */
-void GameObject::SetCapturePointSlider(float value, bool isLocked)
-{
-    GameObjectInfo const* info = GetGOInfo();
 
-    m_captureSlider = value;
 
-    // only activate non-locked capture point
-    if (!isLocked)
-    {
-        SetLootState(GO_ACTIVATED);
-    }
 
-    // set the state of the capture point based on the slider value
-    if ((int)m_captureSlider == CAPTURE_SLIDER_ALLIANCE)
-    {
-        m_captureState = CAPTURE_STATE_WIN_ALLIANCE;
-    }
-    else if ((int)m_captureSlider == CAPTURE_SLIDER_HORDE)
-    {
-        m_captureState = CAPTURE_STATE_WIN_HORDE;
-    }
-    else if (m_captureSlider > CAPTURE_SLIDER_MIDDLE + info->capturePoint.neutralPercent * 0.5f)
-    {
-        m_captureState = CAPTURE_STATE_PROGRESS_ALLIANCE;
-    }
-    else if (m_captureSlider < CAPTURE_SLIDER_MIDDLE - info->capturePoint.neutralPercent * 0.5f)
-    {
-        m_captureState = CAPTURE_STATE_PROGRESS_HORDE;
-    }
-    else
-    {
-        m_captureState = CAPTURE_STATE_NEUTRAL;
-    }
-}
 
-/**
- * @brief Updates capture point progress and sends related world state changes.
- */
-void GameObject::TickCapturePoint()
-{
-    // TODO: On retail: Ticks every 5.2 seconds. slider value increase when new player enters on tick
-
-    GameObjectInfo const* info = GetGOInfo();
-    float radius = info->capturePoint.radius;
-
-    // search for players in radius
-    std::list<Player*> capturingPlayers;
-    MaNGOS::AnyPlayerInCapturePointRange u_check(this, radius);
-    MaNGOS::PlayerListSearcher<MaNGOS::AnyPlayerInCapturePointRange> checker(capturingPlayers, u_check);
-    Cell::VisitWorldObjects(this, checker, radius);
-
-    GuidSet tempUsers(m_UniqueUsers);
-    uint32 neutralPercent = info->capturePoint.neutralPercent;
-    int oldValue = m_captureSlider;
-    int rangePlayers = 0;
-
-    for (std::list<Player*>::iterator itr = capturingPlayers.begin(); itr != capturingPlayers.end(); ++itr)
-    {
-        if ((*itr)->GetTeam() == ALLIANCE)
-        {
-            ++rangePlayers;
-        }
-        else
-        {
-            --rangePlayers;
-        }
-
-        ObjectGuid guid = (*itr)->GetObjectGuid();
-        if (!tempUsers.erase(guid))
-        {
-            // new player entered capture point zone
-            m_UniqueUsers.insert(guid);
-
-            // send capture point enter packets
-            (*itr)->SendUpdateWorldState(info->capturePoint.worldState3, neutralPercent);
-            (*itr)->SendUpdateWorldState(info->capturePoint.worldState2, oldValue);
-            (*itr)->SendUpdateWorldState(info->capturePoint.worldState1, WORLD_STATE_ADD);
-            (*itr)->SendUpdateWorldState(info->capturePoint.worldState2, oldValue); // also redundantly sent on retail to prevent displaying the initial capture direction on client capture slider incorrectly
-        }
-    }
-
-    for (GuidSet::iterator itr = tempUsers.begin(); itr != tempUsers.end(); ++itr)
-    {
-        // send capture point leave packet
-        if (Player* owner = GetMap()->GetPlayer(*itr))
-        {
-            owner->SendUpdateWorldState(info->capturePoint.worldState1, WORLD_STATE_REMOVE);
-        }
-
-        // player left capture point zone
-        m_UniqueUsers.erase(*itr);
-    }
-
-    // return if there are not enough players capturing the point (works because minSuperiority is always 1)
-    if (rangePlayers == 0)
-    {
-        // set to inactive if all players left capture point zone
-        if (m_UniqueUsers.empty())
-        {
-            SetActiveObjectState(false);
-        }
-        return;
-    }
-
-    // prevents unloading gameobject before all players left capture point zone (to prevent m_UniqueUsers not being cleared if grid is set to idle)
-    SetActiveObjectState(true);
-
-    // cap speed
-    int maxSuperiority = info->capturePoint.maxSuperiority;
-    if (rangePlayers > maxSuperiority)
-    {
-        rangePlayers = maxSuperiority;
-    }
-    else if (rangePlayers < -maxSuperiority)
-    {
-        rangePlayers = -maxSuperiority;
-    }
-
-    // time to capture from 0% to 100% is maxTime for minSuperiority amount of players and minTime for maxSuperiority amount of players (linear function: y = dy/dx*x+d)
-    float deltaSlider = info->capturePoint.minTime;
-
-    if (int deltaSuperiority = maxSuperiority - info->capturePoint.minSuperiority)
-    {
-        deltaSlider += (float)(maxSuperiority - abs(rangePlayers)) / deltaSuperiority * (info->capturePoint.maxTime - info->capturePoint.minTime);
-    }
-
-    // calculate changed slider value for a duration of 5 seconds (5 * 100%)
-    deltaSlider = 500.0f / deltaSlider;
-
-    Team progressFaction;
-    if (rangePlayers > 0)
-    {
-        progressFaction = ALLIANCE;
-        m_captureSlider += deltaSlider;
-        if (m_captureSlider > CAPTURE_SLIDER_ALLIANCE)
-        {
-            m_captureSlider = CAPTURE_SLIDER_ALLIANCE;
-        }
-    }
-    else
-    {
-        progressFaction = HORDE;
-        m_captureSlider -= deltaSlider;
-        if (m_captureSlider < CAPTURE_SLIDER_HORDE)
-        {
-            m_captureSlider = CAPTURE_SLIDER_HORDE;
-        }
-    }
-
-    // return if slider did not move a whole percent
-    if ((int)m_captureSlider == oldValue)
-    {
-        return;
-    }
-
-    // on retail this is also sent to newly added players even though they already received a slider value
-    for (std::list<Player*>::iterator itr = capturingPlayers.begin(); itr != capturingPlayers.end(); ++itr)
-    {
-        (*itr)->SendUpdateWorldState(info->capturePoint.worldState2, (uint32)m_captureSlider);
-    }
-
-    // send capture point events
-    uint32 eventId = 0;
-
-    /* WIN EVENTS */
-    // alliance wins tower with max points
-    if (m_captureState != CAPTURE_STATE_WIN_ALLIANCE && (int)m_captureSlider == CAPTURE_SLIDER_ALLIANCE)
-    {
-        eventId = info->capturePoint.winEventID1;
-        m_captureState = CAPTURE_STATE_WIN_ALLIANCE;
-    }
-    // horde wins tower with max points
-    else if (m_captureState != CAPTURE_STATE_WIN_HORDE && (int)m_captureSlider == CAPTURE_SLIDER_HORDE)
-    {
-        eventId = info->capturePoint.winEventID2;
-        m_captureState = CAPTURE_STATE_WIN_HORDE;
-    }
-
-    /* PROGRESS EVENTS */
-    // alliance takes the tower from neutral, contested or horde (if there is no neutral area) to alliance
-    else if (m_captureState != CAPTURE_STATE_PROGRESS_ALLIANCE && m_captureSlider > CAPTURE_SLIDER_MIDDLE + neutralPercent * 0.5f && progressFaction == ALLIANCE)
-    {
-        eventId = info->capturePoint.progressEventID1;
-
-        // handle objective complete
-        if (m_captureState == CAPTURE_STATE_NEUTRAL)
-            if (OutdoorPvP* outdoorPvP = sOutdoorPvPMgr.GetScript((*capturingPlayers.begin())->GetCachedZoneId()))
-            {
-                outdoorPvP->HandleObjectiveComplete(eventId, capturingPlayers, progressFaction);
-            }
-
-        // set capture state to alliance
-        m_captureState = CAPTURE_STATE_PROGRESS_ALLIANCE;
-    }
-    // horde takes the tower from neutral, contested or alliance (if there is no neutral area) to horde
-    else if (m_captureState != CAPTURE_STATE_PROGRESS_HORDE && m_captureSlider < CAPTURE_SLIDER_MIDDLE - neutralPercent * 0.5f && progressFaction == HORDE)
-    {
-        eventId = info->capturePoint.progressEventID2;
-
-        // handle objective complete
-        if (m_captureState == CAPTURE_STATE_NEUTRAL)
-            if (OutdoorPvP* outdoorPvP = sOutdoorPvPMgr.GetScript((*capturingPlayers.begin())->GetCachedZoneId()))
-            {
-                outdoorPvP->HandleObjectiveComplete(eventId, capturingPlayers, progressFaction);
-            }
-
-        // set capture state to horde
-        m_captureState = CAPTURE_STATE_PROGRESS_HORDE;
-    }
-
-    /* NEUTRAL EVENTS */
-    // alliance takes the tower from horde to neutral
-    else if (m_captureState != CAPTURE_STATE_NEUTRAL && m_captureSlider >= CAPTURE_SLIDER_MIDDLE - neutralPercent * 0.5f && m_captureSlider <= CAPTURE_SLIDER_MIDDLE + neutralPercent * 0.5f && progressFaction == ALLIANCE)
-    {
-        eventId = info->capturePoint.neutralEventID1;
-        m_captureState = CAPTURE_STATE_NEUTRAL;
-    }
-    // horde takes the tower from alliance to neutral
-    else if (m_captureState != CAPTURE_STATE_NEUTRAL && m_captureSlider >= CAPTURE_SLIDER_MIDDLE - neutralPercent * 0.5f && m_captureSlider <= CAPTURE_SLIDER_MIDDLE + neutralPercent * 0.5f && progressFaction == HORDE)
-    {
-        eventId = info->capturePoint.neutralEventID2;
-        m_captureState = CAPTURE_STATE_NEUTRAL;
-    }
-
-    /* CONTESTED EVENTS */
-    // alliance attacks tower which is in control or progress by horde (except if alliance also gains control in that case)
-    else if ((m_captureState == CAPTURE_STATE_WIN_HORDE || m_captureState == CAPTURE_STATE_PROGRESS_HORDE) && progressFaction == ALLIANCE)
-    {
-        eventId = info->capturePoint.contestedEventID1;
-        m_captureState = CAPTURE_STATE_CONTEST_HORDE;
-    }
-    // horde attacks tower which is in control or progress by alliance (except if horde also gains control in that case)
-    else if ((m_captureState == CAPTURE_STATE_WIN_ALLIANCE || m_captureState == CAPTURE_STATE_PROGRESS_ALLIANCE) && progressFaction == HORDE)
-    {
-        eventId = info->capturePoint.contestedEventID2;
-        m_captureState = CAPTURE_STATE_CONTEST_ALLIANCE;
-    }
-
-    if (eventId)
-    {
-        StartEvents_Event(GetMap(), eventId, this, this, true, *capturingPlayers.begin());
-    }
-}
-
-// ////////////////////////////////////////////////////////////////////////////////////////////////
-//                              Destructible GO handling
-// ////////////////////////////////////////////////////////////////////////////////////////////////
-void GameObject::DealGameObjectDamage(uint32 damage, uint32 spell, Unit* caster)
-{
-    MANGOS_ASSERT(GetGoType() == GAMEOBJECT_TYPE_DESTRUCTIBLE_BUILDING);
-    MANGOS_ASSERT(spell && sSpellStore.LookupEntry(spell) && caster);
-
-    if (!damage)
-    {
-        return;
-    }
-
-    ForceGameObjectHealth(-int32(damage), caster);
-
-    WorldPacket data(SMSG_DESTRUCTIBLE_BUILDING_DAMAGE, 9 + 9 + 9 + 4 + 4);
-    data << GetPackGUID();
-    data << caster->GetPackGUID();
-    data << caster->GetCharmerOrOwnerOrSelf()->GetPackGUID();
-    data << uint32(damage);
-    data << uint32(spell);
-    SendMessageToSet(&data, false);
-}
-
-void GameObject::RebuildGameObject(Unit* caster)
-{
-    MANGOS_ASSERT(GetGoType() == GAMEOBJECT_TYPE_DESTRUCTIBLE_BUILDING);
-    MANGOS_ASSERT(caster);
-
-    ForceGameObjectHealth(0, caster);
-}
-
-void GameObject::ForceGameObjectHealth(int32 diff, Unit* caster)
-{
-    MANGOS_ASSERT(GetGoType() == GAMEOBJECT_TYPE_DESTRUCTIBLE_BUILDING);
-    MANGOS_ASSERT(caster || diff >= 0);
-
-    if (diff < 0)                                           // Taken damage
-    {
-        DEBUG_FILTER_LOG(LOG_FILTER_DAMAGE, "DestructibleGO: %s taken damage %u dealt by %s", GetGuidStr().c_str(), uint32(-diff), caster->GetGuidStr().c_str());
-#ifdef ENABLE_ELUNA
-        if (caster && caster->ToPlayer())
-        {
-            if (Eluna* e = caster->GetEluna())
-            {
-                e->OnDamaged(this, caster->ToPlayer());
-            }
-        }
-#endif
-        if (m_useTimes > uint32(-diff))
-        {
-            m_useTimes += diff;
-        }
-        else
-        {
-            m_useTimes = 0;
-        }
-    }
-    else if (diff == 0 && GetMaxHealth())                   // Rebuild - TODO: Rebuilding over time with special display-id?
-    {
-        DEBUG_FILTER_LOG(LOG_FILTER_DAMAGE, "DestructibleGO: %s start rebuild by %s", GetGuidStr().c_str(), caster->GetGuidStr().c_str());
-
-        m_useTimes = GetMaxHealth();
-        // Start Event if exist
-        if (caster && m_goInfo->destructibleBuilding.rebuildingEvent)
-        {
-            StartEvents_Event(GetMap(), m_goInfo->destructibleBuilding.rebuildingEvent, this, caster->GetCharmerOrOwnerOrSelf(), true, caster->GetCharmerOrOwnerOrSelf());
-        }
-    }
-    else                                                    // Set to value
-    {
-        m_useTimes = uint32(diff);
-    }
-
-    uint32 newDisplayId = 0xFFFFFFFF;                       // Set to invalid -1 to track if we switched to a change state
-    DestructibleModelDataEntry const* destructibleInfo = sDestructibleModelDataStore.LookupEntry(m_goInfo->destructibleBuilding.destructibleData);
-
-    // Get Current State - Note about order: Important for GetMaxHealth() == 0
-    if (m_useTimes == GetMaxHealth())                       // Full Health
-    {
-        DEBUG_FILTER_LOG(LOG_FILTER_DAMAGE, "DestructibleGO: %s set to full health %u", GetGuidStr().c_str(), m_useTimes);
-
-        RemoveFlag(GAMEOBJECT_FLAGS, GO_FLAG_UNK_9 | GO_FLAG_UNK_10 | GO_FLAG_UNK_11);
-        newDisplayId = m_goInfo->displayId;
-
-        // Start Event if exist
-        if (caster && m_goInfo->destructibleBuilding.intactEvent)
-        {
-            StartEvents_Event(GetMap(), m_goInfo->destructibleBuilding.intactEvent, this, caster->GetCharmerOrOwnerOrSelf(), true, caster->GetCharmerOrOwnerOrSelf());
-        }
-    }
-    else if (m_useTimes == 0)                               // Destroyed
-    {
-        if (!HasFlag(GAMEOBJECT_FLAGS, GO_FLAG_UNK_11))     // Was not destroyed before
-        {
-            DEBUG_FILTER_LOG(LOG_FILTER_DAMAGE, "DestructibleGO: %s got destroyed", GetGuidStr().c_str());
-#ifdef ENABLE_ELUNA
-            if (caster && caster->ToPlayer())
-            {
-                if (Eluna* e = caster->GetEluna())
-                {
-                    e->OnDestroyed(this, caster->ToPlayer());
-                }
-            }
-#endif
-            RemoveFlag(GAMEOBJECT_FLAGS, GO_FLAG_UNK_9 | GO_FLAG_UNK_10);
-            SetFlag(GAMEOBJECT_FLAGS, GO_FLAG_UNK_11);
-
-            // Get destroyed DisplayId
-            if ((!m_goInfo->destructibleBuilding.destroyedDisplayId || m_goInfo->destructibleBuilding.destroyedDisplayId == 1) && destructibleInfo)
-            {
-                newDisplayId = destructibleInfo->destroyedDisplayId;
-            }
-            else
-            {
-                newDisplayId = m_goInfo->destructibleBuilding.destroyedDisplayId;
-            }
-
-            if (!newDisplayId)                              // No proper destroyed display ID exists, fetch damaged
-            {
-                if ((!m_goInfo->destructibleBuilding.damagedDisplayId || m_goInfo->destructibleBuilding.damagedDisplayId == 1) && destructibleInfo)
-                {
-                    newDisplayId = destructibleInfo->damagedDisplayId;
-                }
-                else
-                {
-                    newDisplayId = m_goInfo->destructibleBuilding.damagedDisplayId;
-                }
-            }
-
-            // Start Event if exist
-            if (caster && m_goInfo->destructibleBuilding.destroyedEvent)
-            {
-                StartEvents_Event(GetMap(), m_goInfo->destructibleBuilding.destroyedEvent, this, caster->GetCharmerOrOwnerOrSelf(), true, caster->GetCharmerOrOwnerOrSelf());
-            }
-        }
-    }
-    else if (m_useTimes <= m_goInfo->destructibleBuilding.damagedNumHits) // Damaged
-    {
-        if (!HasFlag(GAMEOBJECT_FLAGS, GO_FLAG_UNK_10))     // Was not damaged before
-        {
-            DEBUG_FILTER_LOG(LOG_FILTER_DAMAGE, "DestructibleGO: %s got damaged (health now %u)", GetGuidStr().c_str(), m_useTimes);
-
-            SetFlag(GAMEOBJECT_FLAGS, GO_FLAG_UNK_10);
-
-            // Get damaged DisplayId
-            if ((!m_goInfo->destructibleBuilding.damagedDisplayId || m_goInfo->destructibleBuilding.damagedDisplayId == 1) && destructibleInfo)
-            {
-                newDisplayId = destructibleInfo->damagedDisplayId;
-            }
-            else
-            {
-                newDisplayId = m_goInfo->destructibleBuilding.damagedDisplayId;
-            }
-
-            // Start Event if exist
-            if (caster && m_goInfo->destructibleBuilding.damagedEvent)
-            {
-                StartEvents_Event(GetMap(), m_goInfo->destructibleBuilding.damagedEvent, this, caster->GetCharmerOrOwnerOrSelf(), true, caster->GetCharmerOrOwnerOrSelf());
-            }
-        }
-    }
-
-    // Set display Id
-    if (newDisplayId != 0xFFFFFFFF && newDisplayId != GetDisplayId() && newDisplayId)
-    {
-        SetDisplayId(newDisplayId);
-    }
-
-    // Set health
-    SetGoAnimProgress(GetMaxHealth() ? m_useTimes * 255 / GetMaxHealth() : 255);
-}
 
 /**
  * @brief Gets the bound script id for this game object.
